// FE/rempms-admin-frontend/src/app/model/userAccount/UserAccount.ts
import {UserRole} from "../userRole/UserRole";

export class UserAccount {
  idUserAccount: string | any;
  userName: string | any;
  password: string | any;
  userImage: string | any;
  hiddenUserImageName: string | any;
  token: string | any;
  commonStatus: string | any;
  idUserRole: string | any;
  userRole: UserRole | any;
}